pet_clinic/requirements.txt
pyodbc==5.3.0
psycopg2-binary==2.9.11
urllib3
requests
